loan_api/requirements.txt
alembic==1.14.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
catboost==1.2.7
email_validator==2.2.0
fastapi==0.115.8
matplotlib==3.10.0
numpy==1.26.4
pandas==2.2.3
passlib==1.7.4
pydantic==2.10.6
python-dotenv==1.0.1
python-jose==3.4.0
scikit-learn==1.6.1
sqlmodel==0.0.22
starlette==0.45.3
uvicorn==0.34.0